Refactor provider structure (#2577)
* refactor: App provider

* Remove app provider

* Refactor userSettingsProvider

* Remove comments

* Add helper to test composables with global providers like user-settings.provider
