+++ src/composables/useNumbers.ts
@@ -1,7 +1,7 @@
 import BigNumber from 'bignumber.js';
 import numeral from 'numeral';
 
-import useUserSettings from '@/composables/useUserSettings';
+import { useUserSettings } from '@/providers/user-settings.provider';
 import { FiatCurrency } from '@/constants/currency';
 import { bnum } from '@/lib/utils';
 import useTokens from './useTokens';
