+++ src/main.ts
@@ -24,6 +24,10 @@ import Web3Plugin from '@/services/web3/web3.plugin';
 import store from '@/store';
 
 import Root from './Root';
+import {
+  userSettingsProvider,
+  UserSettingsProviderSymbol,
+} from './providers/user-settings.provider';
 
 echarts.use([
   TooltipComponent,
@@ -47,6 +51,8 @@ const app = createApp(Root)
 registerDirectives(app);
 registerGlobalComponents(app);
 initSentry(app);
+
+app.provide(UserSettingsProviderSymbol, userSettingsProvider());
 
 app.mount('#app');
